plant_disease_backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
tensorflow==2.20.0
numpy>=1.26.0
Pillow==10.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
httpx==0.25.1
email-validator==2.1.0.post1
